requirements.txt
python-dotenv
pytelegrambotapi
python-docx
PyPDF2
chardet
faiss-cpu
sentence-transformers
torch
numpy
tqdm
ollama
